remove weird --deprecated tot version
